requirements.txt
asgiref==3.7.2
crispy-bootstrap5==2023.10
Django==4.2.6
django-crispy-forms==2.1
django-tinymce==3.6.1
django-widget-tweaks==1.5.0
Pillow==10.1.0
sqlparse==0.4.4
typing_extensions==4.8.0
